+++ next.config.ts
@@ -5,6 +5,13 @@ const devPort = echoConfig.server.ports.dev;
 const prodPort = echoConfig.server.ports.prod;
 
 const nextConfig: NextConfig = {
+  // Allow cross-origin requests for development
+  // @ts-ignore
+  allowedDevOrigins: [
+    "echo.catplot.org", 
+    `localhost:${devPort}`, 
+    `192.168.1.74:${devPort}`
+  ],
   experimental: {
     serverActions: {
       bodySizeLimit: '500mb',